+++ lang/java/avro/src/main/java/org/apache/avro/io/ResolvingDecoder.java
@@ -183,7 +183,7 @@ public class ResolvingDecoder extends ValidatingDecoder {
       return (float) in.readLong();
     } else {
       assert actual == Symbol.FLOAT;
-      return (float) in.readFloat();
+      return in.readFloat();
     }
   }
